refactor: make components for screens in profile
