+++ app/(protected)/(tabs)/profile/index.tsx
@@ -1,28 +1,15 @@
-import {
-  User,
-  Key,
-  LogOut,
-  Home,
-  UserX,
-  AlertTriangle,
-  ChevronRight,
-  Edit,
-} from "lucide-react-native";
-import {
-  Text,
-  View,
-  Alert,
-  Modal,
-  ScrollView,
-  ActivityIndicator,
-  TouchableOpacity,
-} from "react-native";
+import { User, Key, LogOut, Home, UserX } from "lucide-react-native";
+import { Text, View, Alert, ScrollView, ActivityIndicator } from "react-native";
 import { AuthService } from "@/services/server/auth";
 import { router } from "expo-router";
 import { useState, useEffect } from "react";
 import { UserService } from "@/services/server/user";
 import { getUserByAccessToken } from "@/services/utils";
 import type { UserResponse } from "@/types/zod/user";
+import { SettingsItem } from "@/components/ui/tabs/profile/settings-item";
+import { ProfileHeader } from "@/components/ui/tabs/profile/profile-header";
+import { LeaveAccommodationModal } from "@/components/ui/tabs/profile/leave-accommodation-modal";
+import { DeleteAccountModal } from "@/components/ui/tabs/profile/delete-account-modal";
 
 export default function ProfileScreen() {
   const authService = new AuthService();
@@ -111,6 +98,14 @@ export default function ProfileScreen() {
     }
   };
 
+  const handleLeaveAccommodation = () => {
+    if (moderators.length === 1) {
+      cannotLeave();
+    } else {
+      leaveAccommodation();
+    }
+  };
+
   if (loading) {
     return (
       <View className="flex-1 justify-center items-center bg-gray-50">
@@ -125,22 +120,7 @@ export default function ProfileScreen() {
   return (
     <ScrollView className="flex-1 bg-gray-50">
       <View className="px-4 py-6">
-        <View className="items-center mb-8">
-          <View className="bg-cyan-600 rounded-full w-24 h-24 items-center justify-center mb-3">
-            <User size={40} color="white" />
-          </View>
-          <Text className="text-2xl font-bold text-gray-800">
-            {user?.username || "Utilisateur"}
-          </Text>
-          <Text className="text-gray-500">{user?.email || "Aucun email"}</Text>
-          <TouchableOpacity
-            className="flex-row items-center mt-2 bg-cyan-50 px-4 py-2 rounded-full"
-            onPress={() => router.replace(`/profile/edit`)}
-          >
-            <Edit size={16} color="#0891b2" />
-            <Text className="text-cyan-600 ml-1">Modifier le profil</Text>
-          </TouchableOpacity>
-        </View>
+        <ProfileHeader user={user} />
 
         <View className="mb-6">
           <Text className="text-lg font-semibold text-gray-800 mb-2">
@@ -194,108 +174,17 @@ export default function ProfileScreen() {
         </View>
       </View>
 
-      <Modal
-        transparent={true}
+      <LeaveAccommodationModal
         visible={warningLeaveAccommodation}
-        animationType="fade"
-        onRequestClose={() => setWarningLeaveAccommodation(false)}
-      >
-        <View className="flex-1 justify-center items-center bg-black bg-opacity-50">
-          <View className="bg-white p-6 rounded-2xl w-5/6 max-w-md">
-            <View className="items-center mb-4">
-              <View className="bg-amber-100 p-3 rounded-full mb-2">
-                <AlertTriangle size={28} color="#f59e0b" />
-              </View>
-              <Text className="text-xl font-bold text-gray-800">
-                Quitter le colocation
-              </Text>
-            </View>
-            <Text className="text-gray-600 mb-6 text-center">
-              Êtes-vous sûr de vouloir quitter cette colocation ? Vous devrez
-              être invité à nouveau pour le rejoindre.
-            </Text>
-            <View className="flex-row justify-between">
-              <TouchableOpacity
-                className="flex-1 mr-2 bg-gray-200 py-3 rounded-lg items-center"
-                onPress={() => setWarningLeaveAccommodation(false)}
-              >
-                <Text className="font-medium text-gray-800">Annuler</Text>
-              </TouchableOpacity>
-              <TouchableOpacity
-                className="flex-1 ml-2 bg-amber-500 py-3 rounded-lg items-center"
-                onPress={() =>
-                  moderators.length === 1 ? cannotLeave() : leaveAccommodation()
-                }
-              >
-                <Text className="font-medium text-white">Quitter</Text>
-              </TouchableOpacity>
-            </View>
-          </View>
-        </View>
-      </Modal>
+        onClose={() => setWarningLeaveAccommodation(false)}
+        onLeave={handleLeaveAccommodation}
+      />
 
-      <Modal
-        transparent={true}
+      <DeleteAccountModal
         visible={warningDelete}
-        animationType="fade"
-        onRequestClose={() => setWarningDelete(false)}
-      >
-        <View className="flex-1 justify-center items-center bg-black bg-opacity-50">
-          <View className="bg-white p-6 rounded-2xl w-5/6 max-w-md">
-            <View className="items-center mb-4">
-              <View className="bg-red-100 p-3 rounded-full mb-2">
-                <AlertTriangle size={28} color="#ef4444" />
-              </View>
-              <Text className="text-xl font-bold text-gray-800">
-                Supprimer le compte
-              </Text>
-            </View>
-            <Text className="text-gray-600 mb-6 text-center">
-              Cette action ne peut pas être annulée. Toutes vos données seront
-              définitivement supprimées.
-            </Text>
-            <View className="flex-row justify-between">
-              <TouchableOpacity
-                className="flex-1 mr-2 bg-gray-200 py-3 rounded-lg items-center"
-                onPress={() => setWarningDelete(false)}
-              >
-                <Text className="font-medium text-gray-800">Annuler</Text>
-              </TouchableOpacity>
-              <TouchableOpacity
-                className="flex-1 ml-2 bg-red-600 py-3 rounded-lg items-center"
-                onPress={deleteAccount}
-              >
-                <Text className="font-medium text-white">Supprimer</Text>
-              </TouchableOpacity>
-            </View>
-          </View>
-        </View>
-      </Modal>
+        onClose={() => setWarningDelete(false)}
+        onDelete={deleteAccount}
+      />
     </ScrollView>
-  );
-}
-
-function SettingsItem({
-  icon,
-  title,
-  onPress,
-  textColor = "text-gray-800",
-}: {
-  icon: React.ReactNode;
-  title: string;
-  onPress: () => void;
-  textColor?: string;
-}) {
-  return (
-    <TouchableOpacity
-      className="flex-row items-center justify-between px-4 py-4 border-b border-gray-100"
-      onPress={onPress}
-    >
-      <View className="flex-row items-center">
-        {icon}
-        <Text className={`ml-3 font-medium ${textColor}`}>{title}</Text>
-      </View>
-      <ChevronRight size={18} color="#9ca3af" />
-    </TouchableOpacity>
   );
 }
